[C] Fix edge now showing when creating a new node | Some code cleanup of NodePos
This happened because the edge nodes were set to position 0,0 in the graph (as default) and when
the viewport was culling any node not currently in view, so if the viewport was far away from 0,0
the edges would not get drawn.
The edges node position did not matter that much before because their position was dynamically set
based on the position of the child and parent nodes (RequiresSecondPassLayout returned true).